requirements.txt
PyGithub==1.46
docker==4.2.0
semver==2.9.0
pyyaml==5.3
